refactor: move chat-thread to lib/zero/
Move chat-thread module (2 source files) from src/lib/chat-thread/
to src/lib/zero/chat-thread/. This module manages chat threads and
already depends on zero-session-service, making it a natural fit
for the zero/ layer.

Closes #7729
