+++ turbo/apps/web/app/api/internal/callbacks/chat/route.ts
@@ -12,7 +12,7 @@ import {
   generateChatTitle,
   type TitleContextMessage,
 } from "../../../../../src/lib/shared/ai/lightweight-model";
-import { updateChatThreadTitle } from "../../../../../src/lib/chat-thread";
+import { updateChatThreadTitle } from "../../../../../src/lib/zero/chat-thread";
 import type { ChatCallbackPayload } from "../../../../../src/lib/callback/callback-payloads";
 import { logger } from "../../../../../src/lib/logger";
